Created basic project layout.
The project is set up to target main.ml as the project entrypoint.
Dune is added as the build system for this project with yojson as
the explicit library dependency and js_of_ocaml as the implicit
library dependency. The following are targeted as dependencies
for Dune: sprites directory, main.ml, and index.html.