Refactored project layout.
The main source files are separated into three directories: lib,
bin, and test. The bin directory contains main.ml, which serves as
the entrypoint to the project. The lib directory contains the
implementation of the interfaces necessary for this project. The
test directory contains all tests for the interfaces in the lib
directory. In this case, it will be easier to use dune to compile
than ocamlbuild.