Refactored Types into State
Since most of the types and functions in Types are most associated
with the state of the game, it would make sense to move it into the
State module. Moreover, the tick should not be used in the function
to get the next game state. Tick should probably be implemented in
the main game loop. Finally, names of some functions are changed
to reflect a more consistent and succinct style.